finished splitting FixedTemperatureBC and DirichletBC
