--- festim/boundary_conditions/fixed_temperature.py
@@ -1,87 +1,0 @@
-import festim as F
-import ufl
-from dolfinx import fem
-import numpy as np
-
-
-class FixedTemperatureBC:
-    def __init__(self, subdomain, value):
-        self.subdomain = subdomain
-        self.value = value
-
-        self.value_fenics = None
-        self.bc_expr = None
-
-    def define_surface_subdomain_dofs(self, facet_meshtags, mesh, function_space):
-        """Defines the facets and the degrees of freedom of the boundary
-        condition
-
-        Args:
-            facet_meshtags (ddolfinx.mesh.MeshTags): the mesh tags of the
-                surface facets
-            mesh (dolfinx.mesh.Mesh): the mesh
-            function_space (dolfinx.fem.FunctionSpace): the function space
-        """
-        bc_facets = facet_meshtags.find(self.subdomain.id)
-        bc_dofs = fem.locate_dofs_topological(function_space, mesh.fdim, bc_facets)
-
-        return bc_dofs
-
-    def create_value(self, mesh, function_space: fem.FunctionSpace, t: fem.Constant):
-        """Creates the value of the boundary condition as a fenics object and sets it to
-        self.value_fenics.
-        If the value is a constant, it is converted to a fenics.Constant.
-        If the value is a function of t, it is converted to a fenics.Constant.
-        Otherwise, it is converted to a fenics.Function and the
-        expression of the function is stored in self.bc_expr.
-
-        Args:
-            mesh (dolfinx.mesh.Mesh) : the mesh
-            function_space (dolfinx.fem.FunctionSpace): the function space
-            t (dolfinx.fem.Constant): the time
-        """
-        x = ufl.SpatialCoordinate(mesh)
-
-        if isinstance(self.value, (int, float)):
-            self.value_fenics = F.as_fenics_constant(mesh=mesh, value=self.value)
-
-        elif callable(self.value):
-            arguments = self.value.__code__.co_varnames
-
-            if "t" in arguments and "x" not in arguments:
-                # only t is an argument
-                if not isinstance(self.value(t=float(t)), (float, int)):
-                    raise ValueError(
-                        f"self.value should return a float or an int, not {type(self.value(t=float(t)))} "
-                    )
-                self.value_fenics = F.as_fenics_constant(
-                    mesh=mesh, value=self.value(t=float(t))
-                )
-            else:
-                self.value_fenics = fem.Function(function_space)
-                kwargs = {}
-                if "t" in arguments:
-                    kwargs["t"] = t
-                if "x" in arguments:
-                    kwargs["x"] = x
-
-                # store the expression of the boundary condition
-                # to update the value_fenics later
-                self.bc_expr = fem.Expression(
-                    self.value(**kwargs),
-                    function_space.element.interpolation_points(),
-                )
-                self.value_fenics.interpolate(self.bc_expr)
-
-    def update(self, t):
-        """Updates the boundary condition value
-
-        Args:
-            t (float): the time
-        """
-        if callable(self.value):
-            arguments = self.value.__code__.co_varnames
-            if isinstance(self.value_fenics, fem.Constant) and "t" in arguments:
-                self.value_fenics.value = self.value(t=t)
-            else:
-                self.value_fenics.interpolate(self.bc_expr)
